+++ packages/af-utils-react-mobx-table/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @af-utils/react-mobx-table
 
+## 0.0.10
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @af-utils/react-table@0.0.10
+
 ## 0.0.9
 
 ### Patch Changes